requirements.txt
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.6
python-dotenv>=1.0
httpx>=0.27
langchain-core>=0.3.0
langchain-openai>=0.1.0 # 推荐用于对接 DeepSeek
playwright>=1.40.0
pyyaml>=6.0
sse-starlette
